OTV: Factor out hardware states into individual files
This change takes hardware state definitions out of the
broadcast_hardware_machine.go file and puts them into their own files to
improve readability.
