+++ cmd/oceantv/broadcast_hardware_machine.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"time"
 
 	"github.com/ausocean/cloud/cmd/oceantv/registry"
 	"github.com/ausocean/cloud/datastore"
@@ -34,340 +33,6 @@ func newableWithContext(new func(ctx *broadcastContext) any, args ...interface{}
 		return nil, errors.New("init args did not contain required broadcast context")
 	}
 	return new(ctx), nil
-}
-
-type hardwareRestarting struct {
-	stateWithTimeoutFields
-	Substate state
-}
-
-var _ = register(hardwareRestarting{})
-
-func (s hardwareRestarting) Name() string { return "hardwareRestarting" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareRestarting from an existing value.
-func (s hardwareRestarting) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareRestarting(ctx) }, args...)
-}
-
-func newHardwareRestarting(ctx *broadcastContext) *hardwareRestarting {
-	return &hardwareRestarting{newStateWithTimeoutFields(ctx), nil}
-}
-
-// For Marshaling/Unmarshaling.
-type hardwareRestartingStateWrapper struct {
-	Type string          `json:"type"`
-	Data json.RawMessage `json:"data"`
-}
-
-func (s *hardwareRestarting) MarshalJSON() ([]byte, error) {
-	substateType := ""
-	substateData := []byte("null")
-
-	if s.Substate != nil {
-		substateType = s.Substate.(registry.Named).Name()
-		data, err := json.Marshal(s.Substate)
-		if err != nil {
-			return nil, fmt.Errorf("could not marshal substate %s in hardwareRestarting: %w", substateType, err)
-		}
-		substateData = data
-	}
-
-	alias := struct {
-		StateWithTimeoutFields stateWithTimeoutFields         `json:",inline"`
-		Substate               hardwareRestartingStateWrapper `json:"substate"`
-	}{
-		StateWithTimeoutFields: s.stateWithTimeoutFields,
-		Substate: hardwareRestartingStateWrapper{
-			Type: substateType,
-			Data: substateData,
-		},
-	}
-
-	data, err := json.Marshal(alias)
-	if err != nil {
-		return nil, fmt.Errorf("could not marshal alias in hardwareRestarting: %w", err)
-	}
-
-	return data, nil
-}
-
-func (s *hardwareRestarting) UnmarshalJSON(data []byte) error {
-	if s.broadcastContext == nil {
-		return errors.New("hardwareRestarting broadcastContext is nil")
-	}
-
-	alias := struct {
-		StateWithTimeoutFields stateWithTimeoutFields         `json:",inline"`
-		Substate               hardwareRestartingStateWrapper `json:"substate"`
-	}{StateWithTimeoutFields: stateWithTimeoutFields{broadcastContext: s.broadcastContext}}
-
-	if err := json.Unmarshal(data, &alias); err != nil {
-		return fmt.Errorf("could not unmarshal alias in hardwareRestarting: %w", err)
-	}
-
-	s.stateWithTimeoutFields = alias.StateWithTimeoutFields
-
-	// Unmarshal substate.
-	if alias.Substate.Type != "" {
-		substate, err := registry.Get(alias.Substate.Type, s.broadcastContext)
-		if err != nil {
-			return fmt.Errorf("could not get substate from registry for type %s in hardwareRestarting: %w", substate, err)
-		}
-
-		_substate, ok := substate.(state)
-		if !ok {
-			panic(fmt.Sprintf("could not assert substate that should be %s in hardwareRestarting", alias.Substate.Type))
-		}
-
-		if err := json.Unmarshal(alias.Substate.Data, _substate); err != nil {
-			return fmt.Errorf("could not unmarshal data for substate %s in hardwareRestarting: %w", alias.Substate.Type, err)
-		}
-
-		s.Substate = _substate
-	}
-
-	return nil
-}
-
-func (s *hardwareRestarting) enter() {
-	s.LastEntered = time.Now()
-	s.Substate = newHardwareStopping(s.broadcastContext)
-	s.Substate.enter()
-}
-func (s *hardwareRestarting) exit() {}
-func (s *hardwareRestarting) transition() {
-	switch s.Substate.(type) {
-	case *hardwareStopping:
-		s.log("(hardwareRestarting) transitioning from substate hardwareStopping to hardwareStarting")
-		s.Substate.exit()
-		s.Substate = newHardwareStarting(s.broadcastContext)
-		s.Substate.enter()
-	default:
-		panic("hardwareRestarting: unexpected transition")
-	}
-}
-
-func (s *hardwareRestarting) handleTimeEvent(t timeEvent) {
-	switch s.Substate.(type) {
-	case *hardwareStopping:
-		withTimeout := s.Substate.(stateWithTimeout)
-		if withTimeout.timedOut(t.Time) {
-			s.bus.publish(hardwareStopFailedEvent{errors.New("hardware stop timed out")})
-			return
-		}
-
-		s.Substate.(*hardwareStopping).handleTimeEvent(t)
-	case *hardwareStarting:
-		withTimeout := s.Substate.(stateWithTimeout)
-		if withTimeout.timedOut(t.Time) {
-			s.bus.publish(hardwareStartFailedEvent{errors.New("exceeded starting timeout during hardware restart")})
-			return
-		}
-
-		// If the camera is reporting then the start has completed.
-		if s.cameraIsReporting() {
-			s.bus.publish(hardwareStartedEvent{})
-			return
-		}
-	default:
-		// This is unexpected and probably means we haven't saved a substate properly.
-		// So perform a notify log and default to a sensible state.
-		s.logAndNotify(broadcastSoftware, "unexpected substate in hardwareRestarting: %v, re-entering state to initialise substate", s.Substate)
-		s.enter()
-	}
-}
-
-func (s *hardwareRestarting) handleHardwareStoppedEvent(event hardwareStoppedEvent) {
-	s.log("handling hardware stopped event")
-	switch s.Substate.(type) {
-	case *hardwareStopping:
-		s.transition()
-	default:
-		// For any other state ignore.
-	}
-}
-
-func (s *hardwareRestarting) handleHardwareShutdownFailedEvent(event hardwareShutdownFailedEvent) {
-	switch s.Substate.(type) {
-	case *hardwareStopping:
-		s.Substate.(*hardwareStopping).handleHardwareShutdownFailedEvent(event)
-	default:
-		// Ignore.
-	}
-}
-
-func (s *hardwareRestarting) cameraIsReporting() bool {
-	up, err := s.hardware.isUp(s.broadcastContext, model.MacDecode(s.cfg.CameraMac))
-	if err != nil {
-		s.bus.publish(invalidConfigurationEvent{fmt.Errorf("could not get camera reporting status: %w", err)})
-		return false
-	}
-	return up
-}
-
-type hardwareStarting struct {
-	stateWithTimeoutFields
-	*broadcastContext `json:"-"`
-}
-
-var _ = register(hardwareStarting{})
-
-func (s hardwareStarting) Name() string { return "hardwareStarting" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareStarting from an existing value.
-func (s hardwareStarting) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareStarting(ctx) }, args...)
-}
-
-func newHardwareStarting(ctx *broadcastContext) *hardwareStarting {
-	return &hardwareStarting{broadcastContext: ctx}
-}
-
-func (s *hardwareStarting) enter() {
-	s.LastEntered = time.Now()
-	// A MAC of 0 indicates it is invalid or unset, proceed with starting the camera.
-	if s.cfg.ControllerMAC == 0 {
-		s.hardware.start(s.broadcastContext)
-		return
-	}
-
-	// The first check for any known hardware error states.
-	hwErr, err := s.hardware.error(s.broadcastContext)
-	if err != nil {
-		errWrapped := fmt.Errorf("could not get hardware error state: %w", err)
-		s.log(errWrapped.Error())
-		// NOTE here we could do this, however it's not certain that all ESPs
-		// will have the latest firmware that supports this, and so it's not
-		// necessarily a showstopper.
-		// s.bus.publish(invalidConfigurationEvent{errWrapped})
-		// return
-	}
-
-	switch {
-	case errors.Is(hwErr, LowVoltageAlarm):
-		s.log("controller voltage is low, waiting for recovery before starting")
-		s.bus.publish(lowVoltageEvent{})
-		return
-	case errors.Is(hwErr, None):
-		// Continue other checks, this is good.
-	case hwErr != nil:
-		errWrapped := fmt.Errorf("unhandled controller hardware error: %w", hwErr)
-		s.log(errWrapped.Error())
-		s.bus.publish(invalidConfigurationEvent{errWrapped})
-		return
-	default:
-		// This means we failed to get hwErr, which at this stage just means
-		// we have a controller that doesn't have the latest firmware.
-	}
-
-	voltage, err := s.hardware.voltage(s.broadcastContext)
-	if err != nil {
-		errWrapped := fmt.Errorf("could not get hardware voltage: %w", err)
-		s.log(errWrapped.Error())
-		s.bus.publish(invalidConfigurationEvent{errWrapped})
-		return
-	}
-
-	alarmVoltage, err := s.hardware.alarmVoltage(s.broadcastContext)
-	if err != nil {
-		errWrapped := fmt.Errorf("could not get alarm voltage: %w", err)
-		s.log(errWrapped.Error())
-		s.bus.publish(invalidConfigurationEvent{errWrapped})
-		return
-	}
-
-	controllerIsOn, err := s.hardware.isUp(s.broadcastContext, model.MacDecode(s.cfg.ControllerMAC))
-	if err != nil {
-		errWrapped := fmt.Errorf("could not get controller status: %w", err)
-		s.log(errWrapped.Error())
-		s.bus.publish(invalidConfigurationEvent{errWrapped})
-		return
-	}
-
-	if voltage <= alarmVoltage {
-		if controllerIsOn {
-			s.log("voltage less than alarm voltage but controller is on, something is configured incorrectly")
-			s.bus.publish(invalidConfigurationEvent{errors.New("voltage less than alarm voltage but controller is on")})
-			return
-		}
-		s.log("controller voltage is low, waiting for recovery before starting")
-		s.bus.publish(lowVoltageEvent{})
-		return
-	}
-
-	// Not below alarm voltage, but controller is not responding.
-	// This is a critical failure.
-	if !controllerIsOn {
-		s.bus.publish(controllerFailureEvent{errors.New("controller not responding above alarm voltage")})
-		return
-	}
-
-	// Controller is reporting, but we're not above streaming voltage. Need
-	// to wait for recovery.
-	if voltage < s.cfg.RequiredStreamingVoltage {
-		s.log("controller voltage is below required streaming voltage, waiting for recovery before starting")
-		s.bus.publish(lowVoltageEvent{})
-		return
-	}
-
-	// Controller is reporting and we're above streaming voltage, let's power
-	// on the hardware.
-	s.hardware.start(s.broadcastContext)
-}
-
-func (s *hardwareStarting) exit() {}
-
-func (s *hardwareStarting) timedOut(t time.Time) bool {
-	const timeout = 5 * time.Minute
-	if t.Sub(s.LastEntered) > timeout {
-		s.log("timed out starting hardware, last entered: %v, time now: %v", s.LastEntered, t)
-		return true
-	}
-	return false
-}
-
-func (s *hardwareStarting) reset(time.Duration) {}
-
-type hardwareRecoveringVoltage struct {
-	stateFields
-	stateWithTimeoutFields
-}
-
-var _ = register(hardwareRecoveringVoltage{})
-
-func (s hardwareRecoveringVoltage) Name() string { return "hardwareRecoveringVoltage" }
-
-func (s hardwareRecoveringVoltage) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareRecoveringVoltage(ctx) }, args...)
-}
-
-func newHardwareRecoveringVoltage(ctx *broadcastContext) *hardwareRecoveringVoltage {
-	s := newStateWithTimeoutFields(ctx)
-	s.Timeout = time.Duration(sanatisedVoltageRecoveryTimeout(ctx)) * time.Hour
-	return &hardwareRecoveringVoltage{
-		stateWithTimeoutFields: s,
-	}
-}
-
-func (s *hardwareRecoveringVoltage) enter() {
-	s.LastEntered = time.Now()
-}
-
-func sanatisedVoltageRecoveryTimeout(ctx *broadcastContext) int {
-	// If VoltageRecoveryTimeout is not set, default to 4 hours.
-	if ctx.cfg.VoltageRecoveryTimeout == 0 {
-		const defaultRechargeTimeoutHours = 4
-		ctx.log("recharge timeout hours is not set, defaulting to %d", defaultRechargeTimeoutHours)
-		try(
-			ctx.man.Save(nil, func(_cfg *Cfg) { _cfg.VoltageRecoveryTimeout = defaultRechargeTimeoutHours }),
-			"could not save default recharge timeout hours to config",
-			func(msg string, args ...interface{}) { ctx.logAndNotify(broadcastSoftware, msg, args...) },
-		)
-	}
-	return ctx.cfg.VoltageRecoveryTimeout
 }
 
 type hardwareShutdownFailedEvent struct{ error }
@@ -417,29 +82,6 @@ var _ = registerEvent(hardwareShutdownEvent{})
 
 func (e hardwareShutdownEvent) String() string { return "hardwareShutdownEvent" }
 
-type hardwareShuttingDown struct {
-	stateWithTimeoutFields
-}
-
-var _ = register(hardwareShuttingDown{})
-
-func (s hardwareShuttingDown) Name() string { return "hardwareShuttingDown" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareShuttingDown from an existing value.
-func (s hardwareShuttingDown) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareShuttingDown(ctx) }, args...)
-}
-
-func newHardwareShuttingDown(ctx *broadcastContext) *hardwareShuttingDown {
-	return &hardwareShuttingDown{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-func (s *hardwareShuttingDown) enter() {
-	s.LastEntered = time.Now()
-	s.hardware.shutdown(s.broadcastContext)
-}
-func (s *hardwareShuttingDown) exit() {}
-
 type hardwarePowerOffFailedEvent struct{ error }
 
 var _ = registerEvent(hardwarePowerOffFailedEvent{})
@@ -480,293 +122,6 @@ func (e hardwarePowerOffFailedEvent) Is(target error) bool {
 	}
 	return errors.Is(e.error, target)
 }
-
-type hardwarePoweringOff struct {
-	stateWithTimeoutFields
-}
-
-var _ = register(hardwarePoweringOff{})
-
-func (s hardwarePoweringOff) Name() string { return "hardwarePoweringOff" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwarePoweringOff from an existing value.
-func (s hardwarePoweringOff) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwarePoweringOff(ctx) }, args...)
-}
-
-func newHardwarePoweringOff(ctx *broadcastContext) *hardwarePoweringOff {
-	return &hardwarePoweringOff{stateWithTimeoutFields: newStateWithTimeoutFields(ctx)}
-}
-func (s *hardwarePoweringOff) enter() {
-	s.LastEntered = time.Now()
-	s.hardware.stop(s.broadcastContext)
-}
-func (s *hardwarePoweringOff) exit() {}
-
-type hardwareStopping struct {
-	stateWithTimeoutFields
-	Substate state
-}
-
-var _ = register(hardwareStopping{})
-
-func (s hardwareStopping) Name() string { return "hardwareStopping" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareStopping from an existing value.
-func (s hardwareStopping) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareStopping(ctx) }, args...)
-}
-
-func newHardwareStopping(ctx *broadcastContext) *hardwareStopping {
-	return &hardwareStopping{newStateWithTimeoutFields(ctx), nil}
-}
-
-// For Marshaling/Unmarshaling.
-type hardwareStoppingStateWrapper struct {
-	Type string          `json:"type"`
-	Data json.RawMessage `json:"data"`
-}
-
-func (s *hardwareStopping) MarshalJSON() ([]byte, error) {
-	substateType := ""
-	substateData := []byte("null")
-
-	if s.Substate != nil {
-		substateType = s.Substate.(registry.Named).Name()
-		data, err := json.Marshal(s.Substate)
-		if err != nil {
-			return nil, fmt.Errorf("could not marshal substate %s in hardwareStopping: %w", substateType, err)
-		}
-		substateData = data
-	}
-
-	alias := struct {
-		StateWithTimeoutFields stateWithTimeoutFields       `json:",inline"`
-		Substate               hardwareStoppingStateWrapper `json:"substate"`
-	}{
-		StateWithTimeoutFields: s.stateWithTimeoutFields,
-		Substate: hardwareStoppingStateWrapper{
-			Type: substateType,
-			Data: substateData,
-		},
-	}
-
-	data, err := json.Marshal(alias)
-	if err != nil {
-		return nil, fmt.Errorf("could not marshal alias in hardwareStopping: %w", err)
-	}
-
-	return data, nil
-}
-
-func (s *hardwareStopping) UnmarshalJSON(data []byte) error {
-	if s.broadcastContext == nil {
-		return errors.New("hardwareStopping broadcastContext is nil")
-	}
-
-	alias := struct {
-		StateWithTimeoutFields stateWithTimeoutFields       `json:",inline"`
-		Substate               hardwareStoppingStateWrapper `json:"substate"`
-	}{StateWithTimeoutFields: stateWithTimeoutFields{broadcastContext: s.broadcastContext}}
-
-	if err := json.Unmarshal(data, &alias); err != nil {
-		return fmt.Errorf("could not unmarshal data for alias in hardwareStopping: %w", err)
-	}
-
-	s.stateWithTimeoutFields = alias.StateWithTimeoutFields
-
-	// Unmarshal substate.
-	if alias.Substate.Type != "" {
-		substate, err := registry.Get(alias.Substate.Type, s.broadcastContext)
-		if err != nil {
-			return fmt.Errorf("could not get substate from registry for type %s in hardwareStopping: %w", substate, err)
-		}
-
-		_substate, ok := substate.(state)
-		if !ok {
-			panic(fmt.Sprintf("could not assert substate that should be %s in hardwareStopping", alias.Substate.Type))
-		}
-
-		if err := json.Unmarshal(alias.Substate.Data, _substate); err != nil {
-			return fmt.Errorf("could not unmarshal data for substate %s in hardwareStopping: %w", alias.Substate.Type, err)
-		}
-
-		s.Substate = _substate
-	}
-
-	return nil
-}
-
-func (s *hardwareStopping) enter() {
-	s.LastEntered = time.Now()
-	s.Substate = newHardwareShuttingDown(s.broadcastContext)
-	s.Substate.enter()
-}
-func (s *hardwareStopping) exit() {}
-
-func (s *hardwareStopping) transition() {
-	// This should only be called once.
-	switch s.Substate.(type) {
-	case *hardwareShuttingDown:
-		s.log("(hardwareStopping) transitioning from substate hardwareShuttingDown to hardwarePoweringOff")
-		s.Substate.exit()
-		s.Substate = newHardwarePoweringOff(s.broadcastContext)
-		s.Substate.enter()
-	default:
-		panic("hardwareStopping: unexpected transition")
-	}
-}
-
-func (s *hardwareStopping) handleTimeEvent(t timeEvent) {
-	switch s.Substate.(type) {
-	case *hardwareShuttingDown:
-		s.log("(hardwareStopping) handling timeEvent in hardwareStopping state: substate is hardwareShuttingDown")
-		withTimeout := s.Substate.(stateWithTimeout)
-		if withTimeout.timedOut(t.Time) {
-			s.bus.publish(hardwareShutdownFailedEvent{errors.New("hardware shutdown timed out")})
-			return
-		}
-
-		if !s.cameraIsReporting() {
-			s.bus.publish(hardwareShutdownEvent{})
-			s.transition()
-			return
-		}
-		s.log("(hardwareStopping) camera is still reporting, waiting for shutdown to complete")
-
-	case *hardwarePoweringOff:
-		s.log("(hardwareStopping) handling timeEvent in hardwareStopping state: substate is hardwarePoweringOff")
-		withTimeout := s.Substate.(stateWithTimeout)
-		if withTimeout.timedOut(t.Time) {
-			s.bus.publish(hardwarePowerOffFailedEvent{errors.New("hardware power off timed out")})
-			return
-		}
-
-		if !s.cameraIsReporting() {
-			s.bus.publish(hardwareStoppedEvent{})
-			return
-		}
-		s.log("(hardwareStopping) camera is still reporting, waiting for power off to complete")
-	default:
-		// This is unexpected and probably means we haven't saved a substate properly.
-		// So perform a notify log and default to a sensible state.
-		s.logAndNotify(broadcastSoftware, "unexpected substate in hardwareStopping: %v, re-entering state to initialise substate", s.Substate)
-		s.enter()
-	}
-}
-
-func (s *hardwareStopping) handleHardwareShutdownFailedEvent(event hardwareShutdownFailedEvent) {
-	switch s.Substate.(type) {
-	case *hardwareShuttingDown:
-		// We want to get notified for failures and misconfigured configs, and log
-		// when shutdown is skipped.
-		if errors.Is(event, warnSkipShutdown) {
-			s.log("skipping shutdown: %v:", event.Error)
-		} else if errors.Is(event, errNoShutdownActions) {
-			s.logAndNotify(broadcastHardware, "shutdown skipped: %v", event.Error())
-		}
-		s.transition()
-	default:
-		// Ignore.
-	}
-}
-
-func (s *hardwareStopping) handleHardwarePowerOffFailedEvent(event hardwarePowerOffFailedEvent) {
-	switch s.Substate.(type) {
-	case *hardwarePoweringOff:
-		s.bus.publish(hardwareStopFailedEvent{event})
-	default:
-		// Ignore.
-	}
-}
-
-func (s *hardwareStopping) cameraIsReporting() bool {
-	up, err := s.hardware.isUp(s.broadcastContext, model.MacDecode(s.cfg.CameraMac))
-	if err != nil {
-		s.bus.publish(invalidConfigurationEvent{fmt.Errorf("could not get camera reporting status: %w", err)})
-		return false
-	}
-	return up
-}
-
-type hardwareOn struct{}
-
-var _ = register(hardwareOn{})
-
-func (s hardwareOn) Name() string { return "hardwareOn" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareOn from an existing value.
-func (s hardwareOn) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareOn() }, args...)
-}
-
-func newHardwareOn() *hardwareOn { return &hardwareOn{} }
-func (s *hardwareOn) enter()     {}
-func (s *hardwareOn) exit()      {}
-
-type hardwareOff struct{}
-
-var _ = register(hardwareOff{})
-
-func (s hardwareOff) Name() string { return "hardwareOff" }
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareOff from an existing value.
-func (s hardwareOff) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareOff() }, args...)
-}
-
-func newHardwareOff() *hardwareOff { return &hardwareOff{} }
-func (s *hardwareOff) enter()      {}
-func (s *hardwareOff) exit()       {}
-
-type hardwareFailure struct {
-	*broadcastContext `json:"-"`
-	err               error
-}
-
-var _ = register(hardwareFailure{})
-
-func newHardwareFailure(ctx *broadcastContext, err error) *hardwareFailure {
-	return &hardwareFailure{ctx, err}
-}
-
-func (s hardwareFailure) Name() string { return "hardwareFailure" }
-
-func (s hardwareFailure) MarshalJSON() ([]byte, error) {
-	return json.Marshal(struct{ Err string }{Err: s.err.Error()})
-}
-
-func (s *hardwareFailure) UnmarshalJSON(data []byte) error {
-	aux := struct{ Err string }{}
-	if err := json.Unmarshal(data, &aux); err != nil {
-		return err
-	}
-
-	s.err = errors.New(aux.Err)
-	return nil
-}
-
-// New implements registry.Newable for creating a fresh value of
-// hardwareFailure from an existing value.
-func (s hardwareFailure) New(args ...interface{}) (any, error) {
-	return newableWithContext(func(ctx *broadcastContext) any { return newHardwareFailure(ctx, nil) }, args...)
-}
-func (s *hardwareFailure) enter() {
-	notifyMsg := "entering hardware failure state"
-	notifyKind := broadcastGeneric
-	if s.err != nil {
-		if errEvent, ok := s.err.(errorEvent); ok {
-			notifyKind = errEvent.Kind()
-		}
-		notifyMsg = fmt.Sprintf("entering hardware failure state due to: %v", s.err)
-	}
-	s.logAndNotify(notifyKind, "%s", notifyMsg)
-}
-func (s *hardwareFailure) exit() {}
 
 type hardwareStateMachine struct {
 	currentState state
